Adding a new column "isMyAd" in adsTable
Adding a new column in adsTable to store if the current user is the
owner of the ad.
if isMyAd == 1 then the current user created the advertisement
if isMyAd == 0 the ad was read from NFC tag/beam

Note: SQLite do not have a BOOL type. so this column is an INTEGER